feat: Adds livro insertion into db

+++ app/src/main/java/br/ufms/progmobile/caliopelib/entities/Livro.java
@@ -66,7 +66,7 @@ public class Livro implements Serializable {
         return avaliacao;
     }
 
-    public void setAvaliacao(int avaliacao) {
+    public void setAvaliacao(float avaliacao) {
         this.avaliacao = avaliacao;
     }
